+++ src/Request/Capture/RequestFactory.php
@@ -3,6 +3,8 @@
 namespace ArvPayoneApi\Request\Capture;
 
 use ArvPayoneApi\Request\GenericRequestFactory;
+use ArvPayoneApi\Request\Parts\CartFactory;
+use ArvPayoneApi\Request\PaymentTypes;
 use ArvPayoneApi\Request\RequestFactoryContract;
 use ArvPayoneApi\Request\Types;
 
@@ -20,10 +22,15 @@ class RequestFactory implements RequestFactoryContract
         $genericRequest = GenericRequestFactory::create(Types::CAPTURE, $data);
         $context = $data['context'];
 
+        $cart = null;
+        if ($paymentMethod == PaymentTypes::PAYONE_INVOICE_SECURE) {
+            $cart = CartFactory::create($data);
+        }
         return new Capture(
             $genericRequest,
             $referenceId,
-            $context['capturemode']
+            $context['capturemode'],
+            $cart
         );
     }
 }